requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.6
django-appconf==1.0.2
django-compressor==2.0
django-cors-headers==1.1.0
django-custom-user==0.5
django-filter==0.13.0
django-money==0.8
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
gnureadline==6.3.3
gunicorn==19.4.5
ipython==4.1.2
ipython-genutils==0.1.0
jsonfield==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
py-moneyed==0.6.0
python-dateutil==2.5.2
python-memcached==1.57
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.1.0
ua-parser==0.7.0
user-agents==1.0.1
